split code to make it cleaner
